Add repetition and navlinks modules
