+++ index.html
@@ -58,6 +58,6 @@
     <footer class="footer">
         <p>CopyRight</p>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
